Move debugger files around to eliminate notebooks reference
