+++ src/notebooks/debugger/debuggingManagerBase.ts
@@ -14,18 +14,18 @@ import {
     EventEmitter,
     NotebookCell
 } from 'vscode';
-import { IKernel, IKernelProvider } from '../types';
+import { IKernel, IKernelProvider } from '../../kernels/types';
 import { IDisposable } from '../../platform/common/types';
 import { IApplicationShell, ICommandManager, IVSCodeNotebook } from '../../platform/common/application/types';
-import { DebuggingTelemetry } from './constants';
+import { DebuggingTelemetry } from '../../kernels/debugger/constants';
 import { sendTelemetryEvent } from '../../telemetry';
 import { traceError, traceInfoIfCI } from '../../platform/logging';
 import { DataScience } from '../../platform/common/utils/localize';
-import { IKernelDebugAdapterConfig } from './types';
-import { Debugger } from '../../notebooks/debugger/debugger';
-import { KernelDebugAdapterBase } from './kernelDebugAdapterBase';
-import { INotebookControllerManager } from '../../notebooks/types';
-import { IpykernelCheckResult, isUsingIpykernel6OrLater } from '../../notebooks/debugger/helper';
+import { IKernelDebugAdapterConfig } from '../../kernels/debugger/types';
+import { Debugger } from '../../platform/debugger/debugger';
+import { KernelDebugAdapterBase } from '../../kernels/debugger/kernelDebugAdapterBase';
+import { INotebookControllerManager } from '../types';
+import { IpykernelCheckResult, isUsingIpykernel6OrLater } from '../../kernels/debugger/helper';
 
 /**
  * The DebuggingManager maintains the mapping between notebook documents and debug sessions.
