+++ src/notebooks/serviceRegistry.node.ts
@@ -26,6 +26,9 @@ import { IntellisenseProvider } from '../intellisense/intellisenseProvider.node'
 import { RemoteKernelControllerWatcher } from './controllers/remoteKernelControllerWatcher';
 import { ITracebackFormatter } from '../kernels/types';
 import { NotebookTracebackFormatter } from './outputs/tracebackFormatter';
+import { IJupyterDebugService } from '../kernels/debugger/types';
+import { Identifiers } from '../platform/common/constants';
+import { JupyterDebugService } from './debugger/jupyterDebugService.node';
 
 export function registerTypes(serviceManager: IServiceManager) {
     serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, RemoteSwitcher);
@@ -70,4 +73,9 @@ export function registerTypes(serviceManager: IServiceManager) {
         RemoteKernelControllerWatcher
     );
     serviceManager.addSingleton<ITracebackFormatter>(ITracebackFormatter, NotebookTracebackFormatter);
+    serviceManager.addSingleton<IJupyterDebugService>(
+        IJupyterDebugService,
+        JupyterDebugService,
+        Identifiers.RUN_BY_LINE_DEBUGSERVICE
+    );
 }
